restructured the java class files for custom webapp
