+++ Java/Webapp/com/boeing/cloudview/search/ServletInitializer.java
@@ -1,4 +1,4 @@
-package com.exalead.boeing;
+package com.boeing.cloudview.search;
 
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
@@ -7,7 +7,7 @@ public class ServletInitializer extends SpringBootServletInitializer {
 
 	@Override
 	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(DemoApplication.class);
+		return application.sources(CloudViewSearchApplication.class);
 	}
 
 }
